HTML Extend Update + associated css
Edited a number of things to make it so the html pages would extend main.html instead of loading in their own window. I created and linked css files for the forms and associated html files to try to make the extended pages look uniform to the rest of the project. changed urls file because two views had the same 'name' making it confusing despite the code functioning.